requirements.txt
absl-py==1.4.0
datasets==3.2.0
immutabledict==4.2.1
janome==0.5.0
ja_sentence_segmenter==0.0.2
langdetect==1.0.9
nltk==3.9.1
spacy==3.7.5
openai==1.61.0
anthropic==0.45.2
PyYAML==6.0.1